Conflicts:
	client/src/main/java/com/orientechnologies/orient/client/remote/OStorageRemote.java
	commons/src/main/java/com/orientechnologies/common/concur/resource/OSharedResourceIterator.java
	core/src/main/java/com/orientechnologies/orient/core/config/OGlobalConfiguration.java
	core/src/main/java/com/orientechnologies/orient/core/db/document/ODatabaseDocumentTxPooled.java
	core/src/main/java/com/orientechnologies/orient/core/db/graph/OGraphDatabasePooled.java
	core/src/main/java/com/orientechnologies/orient/core/index/OIndexAbstract.java
	core/src/main/java/com/orientechnologies/orient/core/index/OIndexManagerAbstract.java
	core/src/main/java/com/orientechnologies/orient/core/index/OIndexMultiValues.java
	core/src/main/java/com/orientechnologies/orient/core/index/OIndexOneValue.java
	core/src/main/java/com/orientechnologies/orient/core/index/OIndexUnique.java
	core/src/main/java/com/orientechnologies/orient/core/index/hashindex/local/OUniqueHashIndex.java
	core/src/main/java/com/orientechnologies/orient/core/serialization/serializer/record/string/ORecordSerializerJSON.java
	core/src/main/java/com/orientechnologies/orient/core/sql/OCommandExecutorSQLResultsetAbstract.java
	core/src/main/java/com/orientechnologies/orient/core/sql/method/misc/OSQLMethodSize.java
	core/src/main/java/com/orientechnologies/orient/core/storage/OStorageAbstract.java
	core/src/main/java/com/orientechnologies/orient/core/storage/OStorageEmbedded.java
	core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/OStorageLocal.java
	core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/paginated/OLocalPaginatedStorage.java
	core/src/main/java/com/orientechnologies/orient/core/type/tree/OMVRBTreeRID.java
	distributed/config/hazelcast.xml
	distributed/config/orientdb-dserver-config.xml
	distributed/src/main/java/com/orientechnologies/orient/server/hazelcast/OHazelcastPlugin.java
	distributed/src/main/java/com/orientechnologies/orient/server/hazelcast/sharding/hazelcast/OHazelcastDHTNodeProxy.java
	distributed/src/test/java/com/orientechnologies/orient/server/distributed/AbstractServerClusterInsertTest.java
	distributed/src/test/resources/hazelcast-0.xml
	distributed/src/test/resources/hazelcast-1.xml
	distributed/src/test/resources/hazelcast-2.xml
	distributed/src/test/resources/orientdb-dserver-config-0.xml
	distributed/src/test/resources/orientdb-dserver-config-1.xml
	distributed/src/test/resources/orientdb-dserver-config-2.xml
	enterprise/src/main/java/com/orientechnologies/orient/enterprise/channel/OChannel.java
	enterprise/src/main/java/com/orientechnologies/orient/enterprise/channel/binary/OAsynchChannelServiceThread.java
	graphdb/config/orientdb-dserver-config.xml
	graphdb/src/test/java/com/orientechnologies/orient/graph/blueprints/BlueprintsConcurrentAddEdgeTest.java
	graphdb/src/test/java/com/orientechnologies/orient/graph/blueprints/RemoteGremlinTest.java
	object/src/main/java/com/orientechnologies/orient/object/db/ODatabaseObjectTxPooled.java
	object/src/main/java/com/orientechnologies/orient/object/db/OObjectDatabaseTxPooled.java
	pom.xml
	server/src/main/java/com/orientechnologies/orient/server/OServer.java
	server/src/main/java/com/orientechnologies/orient/server/OServerMain.java
	server/src/main/java/com/orientechnologies/orient/server/distributed/ODistributedAbstractPlugin.java
	server/src/main/java/com/orientechnologies/orient/server/distributed/ODistributedServerLog.java
	server/src/main/java/com/orientechnologies/orient/server/distributed/ODistributedStorage.java
	server/src/main/java/com/orientechnologies/orient/server/distributed/task/OAbstractReplicatedTask.java
	server/src/main/java/com/orientechnologies/orient/server/distributed/task/OAlignRequestTask.java
	server/src/main/java/com/orientechnologies/orient/server/distributed/task/OAlignResponseTask.java
	server/src/main/java/com/orientechnologies/orient/server/distributed/task/OReadRecordTask.java
	server/src/main/java/com/orientechnologies/orient/server/distributed/task/OUpdateRecordTask.java
	server/src/main/java/com/orientechnologies/orient/server/journal/ODatabaseJournal.java
	server/src/main/java/com/orientechnologies/orient/server/network/protocol/http/command/OServerCommandAuthenticatedDbAbstract.java
	server/src/site/js/orientdb-api.js
	server/src/site/studio/panelCreateDatabase.htm

+++ core/src/main/java/com/orientechnologies/orient/core/storage/fs/OFile.java
@@ -69,7 +69,7 @@ public interface OFile {
    */
   public abstract void delete() throws IOException;
 
-  public abstract void setSize(int iSize) throws IOException;
+  public abstract void setSize(long iSize) throws IOException;
 
   public abstract void writeHeaderLong(int iPosition, long iValue) throws IOException;
 
@@ -114,18 +114,20 @@ public interface OFile {
   /**
    * Cuts bytes from the tail of the file reducing the filledUpTo size.
    * 
+   * 
    * @param iSizeToShrink
    * @throws IOException
    */
-  public abstract void removeTail(int iSizeToShrink) throws IOException;
+  public abstract void removeTail(long iSizeToShrink) throws IOException;
 
   /**
    * Shrink the file content (filledUpTo attribute only)
    * 
+   * 
    * @param iSize
    * @throws IOException
    */
-  public abstract void shrink(final int iSize) throws IOException;
+  public abstract void shrink(final long iSize) throws IOException;
 
   public abstract String getName();
 
@@ -135,19 +137,19 @@ public interface OFile {
 
   public abstract boolean renameTo(File newFile);
 
-  public abstract int allocateSpace(final int iSize) throws IOException;
+  public abstract long allocateSpace(final long iSize) throws IOException;
 
-  public abstract int getFreeSpace();
+  public abstract long getFreeSpace();
 
-  public abstract int getFileSize();
+  public abstract long getFileSize();
 
-  public abstract int getFilledUpTo();
+  public abstract long getFilledUpTo();
 
   public abstract boolean canOversize(final int iRecordSize);
 
   public abstract String toString();
 
-  public abstract int getMaxSize();
+  public abstract long getMaxSize();
 
   public abstract void setMaxSize(int maxSize);
